Rename extension related config and method
