+++ assets/js/blocks-registry/payment-methods/extensions-config.js
